requirements.txt
aiogram==3.2.0
gspread==5.12.4
python-dotenv==1.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
